requirements.txt
pandas
tiktoken
python-dotenv
openai
tqdm
joblib
scikit-learn